requirements.txt

# BiVAE
cornac>=2.3.5
# Web Framework (FastAPI)
fastapi>=0.104.0
numpy>=1.24.0

# LLM
openai>=1.0.0
pandas>=2.0.0


python-dotenv>=1.0.0
python-multipart>=0.0.6

# Deep Learning (NextItNet)
torch>=2.0.0
uvicorn[standard]>=0.24.0
